gmp: add configure.ac patch to append user provided compilation flags
gmp configure script is pretty good at auto detecting the ABI and the
tune flags that need to be passes to the compiler. However, the user
provided flags (CFLAGS, CXXFLAGS, CPPFLAGS) take precedence and the ABI
detection may fail, leading to configure errors like the one below:

| configure: error: Oops, mp_limb_t is 32 bits, but the assembler code
| in this configuration expects 64 bits.
| You appear to have set $CFLAGS, perhaps you also need to tell GMP the
| intended ABI, see "ABI and ISA" in the manual.

One solution would be to change the recipe and add the ABI manually, or
let gmp do the job.

So, this patch will:
 * allow the configure process to auto-detect the ABI and tune flags
   properly;
 * append our flags to the detected ones;

[YOCTO #5783]

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>