icu: Disable the default LDFLAGSICUDT for target compile
By default LDFLAGSICUDT=-nodefaultlibs -nostdlib for Linux
which means DT_NEEDED section for libicu will not be populated
with dependencies when we reset it to be empty then the default
libraries it needs are added to DT_NEEDED section in ELF header

This patch is a workaround as I consider it for now, since it could
be a problem in glibc dynamic loader for arm (especially for hf case)
where its unable to load shared objects which dont have any dependencies
expressed in DT_NEEDED segment.

here is when LDFLAGSICUDT=-nodefaultlibs -nostdlib

Dynamic section at offset 0x1549c10 contains 8 entries:
  Tag        Type                         Name/Value
 0x0000000e (SONAME)                     Library soname:
[libicudata.so.51]
 0x00000010 (SYMBOLIC)                   0x0
 0x6ffffef5 (GNU_HASH)                   0xf8
 0x00000005 (STRTAB)                     0x188
 0x00000006 (SYMTAB)                     0x138
 0x0000000a (STRSZ)                      54 (bytes)
 0x0000000b (SYMENT)                     16 (bytes)
 0x00000000 (NULL)                       0x0

here is one with LDFLAGSICUDT empty

Dynamic section at offset 0x154a014 contains 22 entries:
  Tag        Type                         Name/Value
 0x00000001 (NEEDED)                     Shared library: [libc.so.6]
 0x0000000e (SONAME)                     Library soname:
[libicudata.so.51]
 0x00000010 (SYMBOLIC)                   0x0
 0x0000000c (INIT)                       0x33c
 0x0000000d (FINI)                       0x500
 0x6ffffef5 (GNU_HASH)                   0xf8
 0x00000005 (STRTAB)                     0x204
 0x00000006 (SYMTAB)                     0x144
 0x0000000a (STRSZ)                      192 (bytes)
 0x0000000b (SYMENT)                     16 (bytes)
 0x00000003 (PLTGOT)                     0x154a0f8
 0x00000002 (PLTRELSZ)                   16 (bytes)
 0x00000014 (PLTREL)                     REL
 0x00000017 (JMPREL)                     0x32c
 0x00000011 (REL)                        0x2fc
 0x00000012 (RELSZ)                      48 (bytes)
 0x00000013 (RELENT)                     8 (bytes)
 0x6ffffffe (VERNEED)                    0x2dc
 0x6fffffff (VERNEEDNUM)                 1
 0x6ffffff0 (VERSYM)                     0x2c4
 0x6ffffffa (RELCOUNT)                   1
 0x00000000 (NULL)                       0x0

btw. ldd reveals the fist one to be static library while the second
one is detected as shared library. This could be a clue into how
elf headers are being interpreted by dynamic loader.

The data seems to be all static in libicudata which could
load it quicker and thats what could be confusing dynamic linker ..may
be

Signed-off-by: Khem Raj <raj.khem@gmail.com>