ltp: Work around parallel make race
As per the comments, the makefiles use make -C extensively and
this causes particular problems around the kernel syscall header. We
therefore ensure its up to date in advance.

Also, append the configure, not prepend since otherwise the configure
will rerun during the build due to the timestamps being outdated.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>