mips64-linux: set ac_cv_sizeof_ssize_t
Set ac_cv_sizeof_ssize_t for mips64;
mips-common will not overwrite it.

"ssize_t is a posix define which is architecture specific whose value
is signed size_t, glibc/uclibc for mips64/n64 linux platform defines
it to be equivalent of 'long' and long here is 8bytes because
mips64/n64 follows LP64 model. In OpenEmbedded our default ABI for
mips64 platforms is N64, having said that autoconf decides to poke at
the platform for finding these kind of sizes which fails when you are
doing cross compiling hence we have to cache it." - Khem Raj

[YOCTO #5935]

Signed-off-by: Valentin Popa <valentin.popa@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>