expect: fix do_install failure on SLED 11.2
Previously we were encountering do_install failures on older machines
like SLED 11.2. These machines have relatively lower versions of glibc.

     tclsh8.6: /lib64/libc.so.6: version `GLIBC_2.14' not found

The above failure is due to that fact that we are using the fixline1
tcl script to install commands under the example directory. The purpose
of fixline1 is to ensure that the installed scripts begin with '#!/bin/sh'.
However, the scripts under the example directory have already got
the correct first line. That's why we don't need to modify anything and could
just copy them.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>