pixbufcache: Fix librsvg-native build
bitbake librsvg-native would fail with an error about missing icu-native.
The reason is that bitbake doesn't directly parse setscene dependencies. This
change ensures bitbake does see the dependencies and avoids the error.

Ideally we'd teach bitbake about those but that is a significant and complex
change so this resolves the problem for now.

[YOCTO #5926]

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>