cmake-native: Stop building ccmake
The bootstrapped cmake is called in such a way that it will automatically
enable building ccmake if curses is found. This tool isn't particularly
useful to us and it will cause build problems if ncurses-native is built in
parallel with cmake-native so let's just pass -DBUILD_CursesDialog=0 to
disable the feature altogether as the non-native cmake does.

Unfortunately this requires patching the bootstrap script since there
appears to be no way to get this option through.

Signed-off-by: Mike Crowe <mac@mcrowe.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>