utils: Add hardlinkdir shell function
In a number of places it would be helpful to be able to copy trees of
files using hardlinks. This turns out to be harder than you'd expect
since there is no good single command that does this well and handles
all file types correctly.

Abstracting this into a function therefore makes sense, cpio seems
as good an option as any other.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>